send temperature to Smartphone

+++ Android/src/ihealth/arduino/MessageReceiver.java
@@ -2,5 +2,5 @@ package ihealth.arduino;
 
 public interface MessageReceiver {
 
-	void receiveMeasurementResult(int value);
+	void receiveMeasurementResult(float value);
 }
